Fix reported column number in ESLint
ESLint's column numbers are 0-based:
https://eslint.org/docs/developer-guide/working-with-rules#contextreport

> column - the 0-based column number at which the problem occurred.

Flow's column numbers are 1-based:

    $ cat index.js
    // @flow
    // The next line has an error at the first column: "abc" is not defined.
    abc

    $ flow check-contents < index.js --json --root index.js | jq '.errors[] | .message[] | .loc'
    {
      "source": "-",
      "type": "SourceFile",
      "start": {
        "line": 3,
        "column": 1,
        "offset": 82
      },
      "end": {
        "line": 3,
        "column": 3,
        "offset": 85
      }
    }

This commit simply subtracts the start column by 1.

Before:

    $ eslint index.js --format codeframe
    error: identifier `abc`:  Could not resolve name (flowtype-errors/show-errors) at index.js:3:2:
      1 | // @flow
      2 | // The next line has an error at the first column: "abc" is not defined.
    > 3 | abc
        |  ^
      4 |

    1 error found.

After:

    $ eslint index.js --format codeframe
    error: identifier `abc`:  Could not resolve name (flowtype-errors/show-errors) at index.js:3:1:
      1 | // @flow
      2 | // The next line has an error at the first column: "abc" is not defined.
    > 3 | abc
        | ^
      4 |

    1 error found.

Shouldn't the end column be subtracted by 1 as well? It seems like
ESLint's end column is exclusive, while Flow's is inclusive. So the
end column should stay unchanged. I tested this by using the
linter-eslint plugin in the Atom editor. When subtracting 1 from the end
column only "ab" of "abc" in index.js above is underlined as an error.
When keeping the end column the entire word "abc" is underlined (which
is the desired result).

I have consistently noted that the error locations are off-by-one after
having used eslint-plugin-flowtype-errors for several months, should it
should be safe to always simply subtract 1 from the start column.